Initial Commit
Flow map C file for use in Matlab along with the Matlab script which processes the flow maps and turns them into a sequence of FTLE matrices.

The spacial and temporal bounds must be specified in the Matlab file while the differential equation's vector field must be specified at the top of the C file.